// lib/types.ts
export type JobCreator = {
  uid: string;
  name: string;
  employeeId?: string;
};

export type Job = {
  id: string;
  title: string;
  from: string;
  to:string;
  pickupCoordinates: { lat: number; lng: number };
  vehicleType: 'Truck' | 'Van' | 'Car';
  status: 'Awaiting' | 'In Progress' | 'Completed';
  requirements: {
    location: string;
    vehicleType: string;
    licenseValidity: boolean;
    insuranceValidity: boolean;
  };
  createdBy: JobCreator;
  createdAt?: string; // ISO 8601 string format
  bookingTime?: string; // ISO 8601 string for scheduled time
  customerName?: string;
  paymentMethod?: 'Invoice' | 'Card' | 'Cash';
  paymentStatus?: 'Paid' | 'Unpaid';
};